bugfix(TDI-38710) : Correct Id type in getLeadChanges & getLeadActivity (#711)

+++ components/components-marketo/src/test/java/org/talend/components/marketo/MarketoConstantsTest.java
@@ -36,7 +36,7 @@ public class MarketoConstantsTest {
     @Test
     public void testLeadChangesSchemaSOAP() throws Exception {
         Schema ref = MarketoConstants.getSOAPSchemaForGetLeadChanges();
-        assertEquals("INT", ref.getField("Id").schema().getTypes().get(0).getType().toString());
+        assertEquals("LONG", ref.getField("Id").schema().getTypes().get(0).getType().toString());
         assertEquals("true", ref.getField("Id").getProp(SchemaConstants.TALEND_COLUMN_IS_KEY));
         assertEquals("LONG", ref.getField("ActivityDateTime").schema().getType().toString());
         assertEquals("java.util.Date", ref.getField("ActivityDateTime").getProp(SchemaConstants.JAVA_CLASS_FLAG));
@@ -51,7 +51,7 @@ public class MarketoConstantsTest {
     @Test
     public void testLeadActivitySchemaSOAP() throws Exception {
         Schema ref = MarketoConstants.getSOAPSchemaForGetLeadActivity();
-        assertEquals("INT", ref.getField("Id").schema().getTypes().get(0).getType().toString());
+        assertEquals("LONG", ref.getField("Id").schema().getTypes().get(0).getType().toString());
         assertEquals("true", ref.getField("Id").getProp(SchemaConstants.TALEND_COLUMN_IS_KEY));
         assertEquals("LONG", ref.getField("ActivityDateTime").schema().getType().toString());
         assertEquals("java.util.Date", ref.getField("ActivityDateTime").getProp(SchemaConstants.JAVA_CLASS_FLAG));
